door: introduce crude setup for entry server setup
This is front-end server of my network. This server is headscale control
panel (VPN controller) and DNS server.

This way even if my internal services are down, I'm always able to
connect to my VPN network and resolve domains to other services.

But if this server is down, I won't be able to connect to anything.

There's still missing fallback DNS setup.